Downgrade timestamps based on our progress through the part
In the new fetch source, we end up with one capability per part, and now
we downgrade that capability as we progress through the file. This means
the overall frontier will progress once we've received all the parts for
a particular time, and all of those parts have made at least some
progress.