persist: enable pushdown feature flags in CI
Both of these flags are "additive" in the sense that the "off" state
does nothing, as opposed to executing some other codepath. This gets us
performance coverage of the new code (e.g. in feature benchmarks).
Combined with the "audit" feature, which we already enabled previously
in CI, we automatically get correctness coverage of both the stats and
interpret paths for anything already covered by existing CI tests.

This is no substitute, of course, for dedicated test coverage, but we
don't lose anything by CI not exercising the "off" case and we gain a
lot. Concretely, this is exactly how we found the bug fixed in #19128.