Make all DatumColumnEncoders nullable
Currently, the type of the field we generate depends on the specific
contents of the array. This means that two arrays generated by the same
encoder may not have the same type, which limits our ability to eg.
concatenate / consolidate them together later.

We don't have nullability information here, and anyway this stuff is
effectively advisory, so there's no harm in making the conservative
choice here.